requirements.txt
accelerate==0.22.0
anytree==2.12.1
bigbench.egg==info
datasets==2.13.1
deepspeed==0.9.5
jsonlines==3.1.0
langchain==0.0.350
nltk==3.8.1
numpy==1.22.2
openai==0.27.8
python_dateutil==2.8.2
ray==2.5.1
Requests==2.31.0
retry==0.9.2
tiktoken==0.4.0
torch==2.0.0a0+1767026
tqdm==4.65.0
transformers==4.33.3
